+++ MoltenVK/MoltenVK/GPUObjects/MVKQueue.h
@@ -180,8 +180,6 @@ protected:
 	friend class MVKQueue;
 
 	MVKQueue* _queue;
-	MVKQueueSubmission* _prev;
-	MVKQueueSubmission* _next;
 	MVKVectorInline<MVKSemaphore*, 8> _waitSemaphores;
 	bool _isAwaitingSemaphores;
 	bool _trackPerformance;