+++ Docs/Whats_New.md
@@ -20,7 +20,7 @@ Released TBD
 
 - Revert to supporting host-coherent memory for linear images on macOS.
 - Ensure Vulkan loader magic number is set every time before returning any dispatchable Vulkan handle.
-
+- Consolidate the various linkable objects into a `MVKLinkableMixin` template base class.
 
 
 MoltenVK 1.0.36